+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a changelog](https://github.com/olivierlacan/keep-a-changelog).
 
  <!--next-version-placeholder-->
+## 0.1.3 (2021-12-09)
+## fixed
+ - issue [#5](../../issues/5) inverted value for status en count.
+ - issue [#6](../../issues/6) source host "any" in conjonction with tcp6 protocol ever mismatched.
+### changes
+ - add command line flag to enable trace (debug level) for all connections check.
+ - extend --dry-mode flag to play a check, display the results to stdout then exit.
+
 ## 0.1.2 (2021-12-03)
 ### fixed
 - issues [#1](../../issues/1), [#2](../../issues/2), [#3](../../issues/3), [#4](../../issues/4)